Extract NotebookListView
